apps/api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
psycopg[binary]==3.2.1
python-jose==3.3.0
passlib==1.7.4
email-validator==2.2.0
langgraph==0.2.35
langchain==0.3.0
chromadb==0.5.5
crawl4ai==0.4.247
litellm==1.51.0
requests==2.32.3
beautifulsoup4==4.12.3
python-dateutil==2.9.0.post0
pytest==8.3.3
python-docx==1.1.2
fpdf2==2.7.9
docling
